+++ changelog.md
@@ -2,3 +2,9 @@
 
 ## Functions
 - ✏️ Added `ae.getActivePanelName()`
+
+## PremiereRemote
+- ✏️ Added `resetSelection()`
+
+### 📝 `UXP`
+- ✅ Fixed some actions not using `await`